requirements.txt
requests==2.26.0
apache-airflow==2.1.2
numpy==1.21.2
pandas==1.3.3
mysql-connector-python==8.0.27
pyspark==3.1.2